Working refactored write quorum
