+++ creator-node/src/routes/files.js
@@ -425,7 +425,6 @@ module.exports = function (app) {
 
     // Record image file entries in DB
     const transaction = await models.sequelize.transaction()
-    let newDataRecord
     try {
       // Record dir file entry in DB
       const createDirFileQueryObj = {
@@ -434,7 +433,7 @@ module.exports = function (app) {
         storagePath: resizeResp.dir.dirDestPath,
         type: 'dir' // TODO - replace with models enum
       }
-      newDataRecord = await DBManager.createNewDataRecord(createDirFileQueryObj, cnodeUserUUID, models.File, transaction)
+      await DBManager.createNewDataRecord(createDirFileQueryObj, cnodeUserUUID, models.File, transaction)
 
       // Record all image res file entries in DB
       // Must be written sequentially to ensure clock values are correctly incremented and populated
@@ -447,7 +446,7 @@ module.exports = function (app) {
           dirMultihash: dirCID,
           fileName: file.sourceFile.split('/').slice(-1)[0]
         }
-        newDataRecord = await DBManager.createNewDataRecord(createImageFileQueryObj, cnodeUserUUID, models.File, transaction)
+        await DBManager.createNewDataRecord(createImageFileQueryObj, cnodeUserUUID, models.File, transaction)
       }
 
       req.logger.info(`route time = ${Date.now() - routestart}`)
@@ -457,9 +456,8 @@ module.exports = function (app) {
       return errorResponseServerError(e)
     }
 
-    const primaryClockVal = newDataRecord.clock
     // Must be awaitted and cannot be try-catched, ensuring that error from inside this rejects request
-    await triggerAndWaitForSecondarySyncs(req, { primaryClockVal, immediate: true, enforceQuorum: false })
+    await triggerAndWaitForSecondarySyncs(req, { enforceQuorum: false })
 
     return successResponse({ dirCID })
   }))
